Locking for Tiles
I'm basing this code on the gsoc2009 branch.
The changes create 4 locks in GeglTile, read and write for cpu and opencl, instead of
the old general one.
this is necessary because we can't guarantee that reading is in sync anyway, because
we can change data in the gpu without updating the correspondent data in cpu (and vice-versa).